+++ CHANGELOG.md
@@ -7,12 +7,16 @@ The format is based on [Keep a Changelog][changelog] and this project adheres to
 
 ## Unreleased
 
+### Changed
+
+- The error message when failing to restore parts of a snapshot
+
 ## 11.4.1 - 2021-08-09
 
 ### Fixed
 
 - The order of operations when performing a `copy --mirror` command to avoid
-  snapshot corruption if the operation is cancelled
+  snapshot corruption if the operation is canceled
 
 ## 11.4.0 - 2021-08-09